Add `buildMeta` helper namespace
Add a `buildMeta` global namespace which contains a number of helper
functions which can be used to compute version numbers of builds.

The main useful function is `buildMeta.version()` which reproduces the
code in h's `_version.py`, calculating version numbers for builds as

    <YYYYMMDD>-g<commitsha>[.dirty]

where the date corresponds to the date of the commit at HEAD rather than
the date when the build is running (although these will usually be the
same).

This means that for the time being we are computing these versions
independently in the build and in the applications, but in future it may
be possible (and desirable) to remove the Python `_version.py` helpers
altogether by feeding this version number into the Docker build. For now
this helper shouldn't change anything from the perspective of the Python
applications.