Updated deployApp to support multiple regions
deployApp.groovy has been modified to include a region parameter which
is then forwarded onto the call it makes to the deployment workflow. The
value can be NULL and does not need to be passed by what ever is calling
deployApp. In such a scenario, the value of region will be set to
`us-west-1`.